Responding to Requirements
As RealTerm has now been on v2.x for 4 years now, perhaps v3 is a long way off! I've adapted this and the chocolateyinstall.ps1 script to try to conform to the requirements on the message on https://chocolatey.org/packages/realterm. I noted your error, the author has responded to a similar error here, what do you think?